create DeliveryStatus constant
