+++ constants/index.js
@@ -1,0 +1,10 @@
+const DeliveryStatus = {
+  RINGDOWN_SENT: 'RINGDOWN SENT',
+  RINGDOWN_RECEIVED: 'RINGDOWN RECEIVED',
+  ARRIVED: 'ARRIVED',
+  OFFLOADED: 'OFFLOADED',
+  RETURNED_TO_SERVICE: 'RETURNED TO SERVICE',
+};
+Object.freeze(DeliveryStatus);
+
+exports.DeliveryStatus = DeliveryStatus;
